# Speed up entity lookup using a trie