cc @copy

Another FYI PR, in the manner of #15.

This:
- Does not affect the parsing time of the kind of input you linked me to, unfortunately. The source of that page contained at most one entity reference.
- Reduces time to parse the OCaml Reddit by about 20% – from 29 ms to 23 ms in my testing.
- Reduces time to parse `ocamldoc`s of a sizeable OCaml project by about 28 times (!!) – from 14s to 0.5s. `ocamldoc` uses a lot of `&nbsp;` for spacing, and other entities for escaping code blocks.